MyRiversideCleaner/riverside/requirements.txt
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cssselect==1.1.0
cssutils==2.3.0
emails==0.6
Flask==2.0.2
Flask-Mail==0.9.1
Flask-Moment==1.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
premailer==3.10.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2

